+++ LuckParser/ConsoleProgram.cs
@@ -1,4 +1,5 @@
-﻿using LuckParser.Controllers;
+﻿using LuckParser.Builders;
+using LuckParser.Controllers;
 using LuckParser.Exceptions;
 using LuckParser.Models.DataModels;
 using System;
