+++ LuckParser/GeneralHelper.cs
@@ -7,7 +7,7 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using static LuckParser.Models.DataModels.ParseEnum.TrashIDS;
 
-namespace LuckParser.Controllers
+namespace LuckParser
 {
     public static class GeneralHelper
     {
