+++ LuckParser/LuckParser.csproj
@@ -85,14 +85,14 @@
     <Compile Include="Exceptions\SkipException.cs" />
     <Compile Include="Exceptions\CancellationException.cs" />
     <Compile Include="ConsoleProgram.cs" />
-    <Compile Include="Controllers\CSVHelper.cs" />
+    <Compile Include="Builders\HTMLBuilder.cs" />
+    <Compile Include="Builders\CSVBuilder.cs" />
+    <Compile Include="Builders\RawFormatBuilder.cs" />
+    <Compile Include="Builders\Helpers\CombatReplayHelper.cs" />
+    <Compile Include="Builders\Helpers\CSVHelper.cs" />
+    <Compile Include="GeneralHelper.cs" />
     <Compile Include="Controllers\FlashWindow.cs" />
-    <Compile Include="Controllers\GeneralHelper.cs" />
-    <Compile Include="Controllers\RawFormatBuilder.cs" />
-    <Compile Include="Controllers\HTMLBuilder.cs" />
-    <Compile Include="Controllers\CombatReplayHelper.cs" />
     <Compile Include="Controllers\Parser.cs" />
-    <Compile Include="Controllers\CSVBuilder.cs" />
     <Compile Include="Controllers\GW2APIController.cs" />
     <Compile Include="Controllers\ParseHelper.cs" />
     <Compile Include="Controllers\StatisticsCalculator.cs" />
